# docs/reconstruction/assets/openapi.json
{
  "components": {
    "schemas": {
      "ActivityEvent": {
        "additionalProperties": false,
        "properties": {
          "attempt_id": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_:.-]+$",
            "title": "Attempt Id",
            "type": "string"
          },
          "case_ref": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_:.-]+$",
            "title": "Case Ref",
            "type": "string"
          },
          "event_id": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_:.-]+$",
            "title": "Event Id",
            "type": "string"
          },
          "job_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_:.-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "node": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_:.-]+$",
            "title": "Node",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(?:\\.\\d+)?(?:Z|\\+00:00)$",
            "title": "Occurred At",
            "type": "string"
          },
          "operation_id": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_:.-]+$",
            "title": "Operation Id",
            "type": "string"
          },
          "parent_operation_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_:.-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Operation Id"
          },
          "payload": {
            "discriminator": {
              "mapping": {
                "background": "#/components/schemas/BackgroundStatus",
                "model": "#/components/schemas/Lifecycle",
                "narration": "#/components/schemas/Narration",
                "node": "#/components/schemas/Lifecycle",
                "node_summary": "#/components/schemas/NodeSummary",
                "tool": "#/components/schemas/Lifecycle"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/Lifecycle"
              },
              {
                "$ref": "#/components/schemas/NodeSummary"
              },
              {
                "$ref": "#/components/schemas/Narration"
              },
              {
                "$ref": "#/components/schemas/BackgroundStatus"
              }
            ],
            "title": "Payload"
          },
          "run_id": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_:.-]+$",
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "scope": {
            "enum": [
              "CASE",
              "REFUND",
              "MEMORY"
            ],
            "title": "Scope",
            "type": "string"
          },
          "seq": {
            "minimum": 1.0,
            "title": "Seq",
            "type": "integer"
          }
        },
        "required": [
          "event_id",
          "case_ref",
          "run_id",
          "scope",
          "node",
          "operation_id",
          "attempt_id",
          "occurred_at",
          "payload",
          "seq"
        ],
        "title": "ActivityEvent",
        "type": "object"
      },
      "ActivityFacts": {
        "additionalProperties": false,
        "description": "Finite allowlist; raw prose, URLs, messages and documents are excluded.",
        "properties": {
          "action": {
            "anyOf": [
              {
                "enum": [
                  "FULL_REFUND",
                  "DECLINE"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          },
          "count": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count"
          },
          "finding_statuses": {
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_:.-]+$",
              "type": "string"
            },
            "maxItems": 30,
            "title": "Finding Statuses",
            "type": "array"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/ActivityFinding"
            },
            "maxItems": 30,
            "title": "Findings",
            "type": "array"
          },
          "next_node": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_:.-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Node"
          },
          "outcome": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_:.-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "reason_codes": {
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_:.-]+$",
              "type": "string"
            },
            "maxItems": 30,
            "title": "Reason Codes",
            "type": "array"
          },
          "references": {
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_:.-]+$",
              "type": "string"
            },
            "maxItems": 30,
            "title": "References",
            "type": "array"
          },
          "revision_round": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision Round"
          },
          "verdict": {
            "anyOf": [
              {
                "enum": [
                  "APPROVE",
                  "REVISE"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          }
        },
        "title": "ActivityFacts",
        "type": "object"
      },
      "ActivityFinding": {
        "additionalProperties": false,
        "properties": {
          "claim_id": {
            "$ref": "#/components/schemas/ClaimId"
          },
          "evidence_refs": {
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_:.-]+$",
              "type": "string"
            },
            "maxItems": 30,
            "title": "Evidence Refs",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/ClaimStatus"
          }
        },
        "required": [
          "claim_id",
          "status"
        ],
        "title": "ActivityFinding",
        "type": "object"
      },
      "ActivityInputs": {
        "additionalProperties": false,
        "properties": {
          "argument_count": {
            "minimum": 0.0,
            "title": "Argument Count",
            "type": "integer"
          },
          "reason_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReasonCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "required_claim_ids": {
            "items": {
              "$ref": "#/components/schemas/ClaimId"
            },
            "maxItems": 30,
            "title": "Required Claim Ids",
            "type": "array"
          },
          "top_k": {
            "anyOf": [
              {
                "maximum": 3.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K"
          }
        },
        "required": [
          "argument_count"
        ],
        "title": "ActivityInputs",
        "type": "object"
      },
      "ActivityPage": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/ActivityEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "next_cursor": {
            "minimum": 0.0,
            "title": "Next Cursor",
            "type": "integer"
          }
        },
        "required": [
          "events",
          "next_cursor",
          "has_more"
        ],
        "title": "ActivityPage",
        "type": "object"
      },
...
      "ApprovedMemory": {
        "additionalProperties": false,
        "properties": {
          "approved_at": {
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(?:\\.\\d+)?(?:Z|\\+00:00)$",
            "title": "Approved At",
            "type": "string"
          },
          "claim_registry_version": {
            "minLength": 1,
            "title": "Claim Registry Version",
            "type": "string"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "memory_id": {
            "minLength": 1,
            "title": "Memory Id",
            "type": "string"
          },
          "policy_version": {
            "minLength": 1,
            "title": "Policy Version",
            "type": "string"
          },
          "recommended_behavior": {
            "minLength": 1,
            "title": "Recommended Behavior",
            "type": "string"
          },
          "retrieval_summary": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Retrieval Summary",
            "type": "string"
          },
          "scope": {
            "$ref": "#/components/schemas/MemoryScope"
          },
          "status": {
            "const": "APPROVED",
            "title": "Status",
            "type": "string"
          },
          "trigger_conditions": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "minItems": 1,
            "title": "Trigger Conditions",
            "type": "array"
          }
        },
        "required": [
          "memory_id",
          "retrieval_summary",
          "status",
          "recommended_behavior",
          "trigger_conditions",
          "policy_version",
          "claim_registry_version",
          "scope",
          "confidence",
          "approved_at"
        ],
        "title": "ApprovedMemory",
        "type": "object"
      },
...
      "BackgroundStatus": {
        "additionalProperties": false,
        "properties": {
          "error_code": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_:.-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "status": {
            "enum": [
              "SCHEDULED",
              "STARTED",
              "RETRYING",
              "COMPLETED",
              "SKIPPED",
              "FAILED"
            ],
            "title": "Status",
            "type": "string"
          },
          "type": {
            "const": "background",
            "default": "background",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "BackgroundStatus",
        "type": "object"
      },
...
      "ClaimId": {
        "enum": [
          "DELIVERY_CONFIRMED",
          "ORDER_WITHIN_RETURN_WINDOW",
          "SHIPMENT_SEAL_INTACT",
          "ITEM_PHYSICALLY_DAMAGED",
          "DAMAGE_PRESENT_ON_ARRIVAL",
          "ITEM_FUNCTIONALLY_IMPAIRED",
          "ITEM_DIFFERS_FROM_LISTING",
          "WRONG_ITEM_RECEIVED",
          "ITEM_NOT_IN_SHIPMENT",
          "ITEM_UNUSED"
        ],
        "title": "ClaimId",
        "type": "string"
      },
      "ClaimStatus": {
        "enum": [
          "SUPPORTED",
          "UNSUPPORTED",
          "CONTRADICTED"
        ],
        "title": "ClaimStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Lifecycle": {
        "additionalProperties": false,
        "properties": {
          "duration_ms": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "error_code": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_:.-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "facts": {
            "$ref": "#/components/schemas/ActivityFacts"
          },
          "inputs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActivityInputs"
              },
              {
                "type": "null"
              }
            ]
          },
          "model": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_:.-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_:.-]+$",
            "title": "Name",
            "type": "string"
          },
          "phase": {
            "enum": [
              "STARTED",
              "COMPLETED",
              "PAUSED",
              "FAILED"
            ],
            "title": "Phase",
            "type": "string"
          },
          "type": {
            "enum": [
              "node",
              "model",
              "tool"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "phase",
          "name"
        ],
        "title": "Lifecycle",
        "type": "object"
      },
...
      "MemoryRetrievalObservation": {
        "additionalProperties": false,
        "allOf": [
          {
            "else": {
              "properties": {
                "error_code": {
                  "type": "string"
                },
                "hits": {
                  "maxItems": 0
                }
              },
              "required": [
                "error_code"
              ]
            },
            "if": {
              "properties": {
                "status": {
                  "const": "OK"
                }
              },
              "required": [
                "status"
              ]
            },
            "then": {
              "properties": {
                "error_code": {
                  "type": "null"
                },
                "query_summary": {
                  "minLength": 1,
                  "type": "string"
                }
              },
              "required": [
                "query_summary"
              ]
            }
          }
        ],
        "properties": {
          "error_code": {
            "anyOf": [
              {
                "enum": [
                  "SUMMARY_UNAVAILABLE",
                  "RETRIEVAL_UNAVAILABLE"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "hits": {
            "items": {
              "$ref": "#/components/schemas/MemorySearchHit"
            },
            "maxItems": 3,
            "title": "Hits",
            "type": "array"
          },
          "query_summary": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Summary"
          },
          "status": {
            "enum": [
              "OK",
              "UNAVAILABLE"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "MemoryRetrievalObservation",
        "type": "object"
      },
      "MemoryScope": {
        "additionalProperties": false,
        "properties": {
          "categories": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "title": "Categories",
            "type": "array"
          },
          "claim_ids": {
            "items": {
              "$ref": "#/components/schemas/ClaimId"
            },
            "title": "Claim Ids",
            "type": "array"
          },
          "market": {
            "minLength": 1,
            "title": "Market",
            "type": "string"
          },
          "reason_codes": {
            "items": {
              "$ref": "#/components/schemas/ReasonCode"
            },
            "title": "Reason Codes",
            "type": "array"
          }
        },
        "required": [
          "market"
        ],
        "title": "MemoryScope",
        "type": "object"
      },
      "MemorySearchHit": {
        "additionalProperties": false,
        "properties": {
          "memory": {
            "$ref": "#/components/schemas/ApprovedMemory"
          },
          "similarity": {
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Similarity",
            "type": "number"
          }
        },
        "required": [
          "memory",
          "similarity"
        ],
        "title": "MemorySearchHit",
        "type": "object"
      },
...
      "Narration": {
        "additionalProperties": false,
        "properties": {
          "error_code": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9_:.-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "source_event_id": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_:.-]+$",
            "title": "Source Event Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "COMPLETED",
              "UNAVAILABLE"
            ],
            "title": "Status",
            "type": "string"
          },
          "text": {
            "anyOf": [
              {
                "maxLength": 600,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "type": {
            "const": "narration",
            "default": "narration",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "source_event_id",
          "status"
        ],
        "title": "Narration",
        "type": "object"
      },
      "NodeSummary": {
        "additionalProperties": false,
        "properties": {
          "facts": {
            "$ref": "#/components/schemas/ActivityFacts"
          },
          "memory_retrieval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MemoryRetrievalObservation"
              },
              {
                "type": "null"
              }
            ]
          },
          "review_gate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewGateResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "type": {
            "const": "node_summary",
            "default": "node_summary",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "facts"
        ],
        "title": "NodeSummary",
        "type": "object"
      },
...
      "ReasonCode": {
        "enum": [
          "ITEM_DAMAGED",
          "ITEM_NOT_AS_DESCRIBED",
          "MISSING_ITEM",
          "WRONG_ITEM",
          "QUALITY_ISSUE",
          "CHANGED_MIND"
        ],
        "title": "ReasonCode",
        "type": "string"
      },
...
      "ReviewGateResult": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "pattern": "^(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "title": "Amount",
            "type": "string"
          },
          "config_hash": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Config Hash",
            "type": "string"
          },
          "config_version": {
            "minLength": 1,
            "title": "Config Version",
            "type": "string"
          },
          "currency": {
            "maxLength": 3,
            "minLength": 3,
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "enum": [
                  "HIGH_VALUE_ITEM",
                  "CURRENCY_THRESHOLD_UNCONFIGURED"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "enum": [
              "PASS",
              "HUMAN_REQUIRED",
              "NOT_APPLICABLE"
            ],
            "title": "Status",
            "type": "string"
          },
          "threshold": {
            "anyOf": [
              {
                "pattern": "^(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold"
          }
        },
        "required": [
          "config_version",
          "config_hash",
          "status",
          "amount",
          "currency",
          "threshold",
          "reason"
        ],
        "title": "ReviewGateResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Adaptive Return Resolution API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/cases/{case_ref}/activities": {
      "get": {
        "operationId": "get_activities_cases__case_ref__activities_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_ref",
            "required": true,
            "schema": {
              "title": "Case Ref",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_seq",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After Seq",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Activities"
      }
    },
    "/cases/{case_ref}/activities/stream": {
      "get": {
        "operationId": "activity_stream_cases__case_ref__activities_stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_ref",
            "required": true,
            "schema": {
              "title": "Case Ref",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_seq",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After Seq",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activity Stream"
      }
    },
...
  }
}